requirements.txt
albumentations==1.3.1
asgiref==3.7.2
attrs==23.1.0
backgroundremover==0.2.5
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cmake==3.27.7
coloredlogs==15.0.1
commandlines==0.4.1
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.3
decorator==4.4.2
Django==4.2.6
dlib==19.24.2
easydict==1.10
face-recognition==1.3.0
face-recognition-models==0.3.0
ffmpeg-python==0.2.0
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.43.1
fsspec==2023.10.0
future==0.18.3
gunicorn==22.0.0
hsh==1.1.0
humanfriendly==10.0
idna==3.4
imageio==2.31.5
imageio-ffmpeg==0.4.9
importlib-resources==6.1.0
imutils==0.5.4
insightface==0.7.3
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lazy_loader==0.3
llvmlite==0.41.1
MarkupSafe==2.1.3
matplotlib==3.8.0
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpy==1.26.0
onnx==1.14.1
onnxruntime==1.16.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
Pillow==9.5.0
platformdirs==3.11.0
pooch==1.8.0
prettytable==3.9.0
proglog==0.1.10
protobuf==4.24.4
PyMatting==1.1.11
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
PyYAML==6.0.1
qudida==0.0.4
referencing==0.30.2
rembg==2.0.52
requests==2.31.0
rpds-py==0.12.0
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
sqlparse==0.4.4
sympy==1.12
threadpoolctl==3.2.0
tifffile==2023.9.26
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==1.26.6
waitress==2.1.2
wcwidth==0.2.8
whitenoise==6.6.0
zipp==3.17.0
